+++ agora-scripts/Codec/Serialise/Orphans.hs
@@ -9,12 +9,10 @@ Orphan instances for Serialising and Hashing Cardano types.
 -}
 module Codec.Serialise.Orphans () where
 
-import Codec.Serialise (Serialise, serialise)
-import Data.Hashable (Hashable (hashWithSalt))
+import Codec.Serialise (Serialise)
 import Data.Tagged (Tagged (Tagged))
 import PlutusLedgerApi.V1 (TxId, TxOutRef)
 import PlutusLedgerApi.V1.Value (AssetClass, CurrencySymbol, TokenName)
-import Types (ScriptParams)
 
 deriving anyclass instance
   Serialise TxOutRef
@@ -34,8 +32,3 @@ deriving anyclass instance
 deriving newtype instance
   Serialise a =>
   Serialise (Tagged s a)
-
-deriving anyclass instance Serialise ScriptParams
-
-instance Hashable ScriptParams where
-  hashWithSalt s scriptParams = hashWithSalt s (serialise scriptParams)
