--- agora-scripts/agora-params.json
@@ -1,11 +1,0 @@
-{
-  "governorInitialSpend": {
-    "txOutRefId": "7be688c61c209dd7a9f4948090db0b031b11a8850b0ee4695f786fea85fbfdee",
-    "txOutRefIdx": 0
-  },
-  "gtClassRef": [
-    "",
-    ""
-  ],
-  "maximumCosigners": 5
-}
